% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chnkScobiInput.R
\docType{data}
\name{chnkScobiInput}
\alias{chnkScobiInput}
\title{Chinook Adult Data from LGTrappingDB formatted for SCOBI}
\format{Data is formatted and ready for \code{SCOBI()}. Originally saved as a csv file}
\usage{
chnkScobiInput
}
\description{
This is an example dataset for adult Chinook from SY2015 that has already been formatted for \code{SCOBI()}.
}
\keyword{Chinook}
\keyword{adult}
\keyword{data}

