% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chnkScrapiInput.R
\docType{data}
\name{chnkScrapiInput}
\alias{chnkScrapiInput}
\title{Chinook Smolt Data from LGTrappingDB formatted for SCRAPI}
\format{Data is formatted and ready for \code{SCRAPI()}. Originally saved as a csv file}
\usage{
chnkScrapiInput
}
\description{
This is a Chinook smolt data set from MY2015 that has already been formatted for \code{SCRAPI()}.
}
\keyword{Chinook}
\keyword{data}
\keyword{smolt}

