% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chnkSmoltPassData.R
\docType{data}
\name{chnkSmoltPassData}
\alias{chnkSmoltPassData}
\title{MY2015 Chinook Smolt Passage Data}
\format{Data is formatted and ready for \code{SCRAPI()}. Originally saved as a csv file}
\usage{
chnkSmoltPassData
}
\description{
This is a Chinook smolt passage data set from MY2015 that has already been formatted for \code{SCRAPI()}.
}
\keyword{Chinook}
\keyword{data}
\keyword{passage}
\keyword{smolt}

