% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chnkWindowCounts.R
\docType{data}
\name{chnkWindowCounts}
\alias{chnkWindowCounts}
\title{SY2015 Chinook Adult Window Count Data}
\format{Data is formatted and ready for \code{SCOBI()}. Originally saved as a csv file}
\usage{
chnkWindowCounts
}
\description{
This is a Chinook window count data set from SY2015 that has already been formatted for \code{SCOBI()}.
}
\keyword{Chinook}
\keyword{adult}
\keyword{data}
\keyword{window}

