% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exRawChnkAdultData.R
\docType{data}
\name{exRawChnkAdultData}
\alias{exRawChnkAdultData}
\title{Example Chinook Adult Data from LGTrappingDB}
\format{Data exported from the LGTrappingDB and originally saved as a csv file}
\usage{
exRawChnkAdultData
}
\description{
This is raw Chinook adult data 'dumped' out of the LGTrappingDB that was originally saved as a .csv file. No formatting
has been done whatsoever.
}
\keyword{Chinook}
\keyword{adult}
\keyword{data}

