% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exRawChnkSmoltData.R
\docType{data}
\name{exRawChnkSmoltData}
\alias{exRawChnkSmoltData}
\title{Example Chinook Yearling Smolt Data from LGTrappingDB}
\format{Data exported from the LGTrappingDB and originally saved as a csv file}
\usage{
exRawChnkSmoltData
}
\description{
This is raw Chinook yearling smolt data 'dumped' out of the LGTrappingDB that was originally saved as a .csv file. No formatting
has been done whatsoever.
}
\keyword{Chinook}
\keyword{data}
\keyword{smolt}

