% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exRawSthdAdultData.R
\docType{data}
\name{exRawSthdAdultData}
\alias{exRawSthdAdultData}
\title{Example Steelhead Adult Data from LGTrappingDB}
\format{Data exported from the LGTrappingDB and originally saved as a csv file}
\usage{
exRawSthdAdultData
}
\description{
This is raw steelhead adult data 'dumped' out of the LGTrappingDB that was originally saved as a .csv file. No formatting
has been done whatsoever.
}
\keyword{adult}
\keyword{data}
\keyword{steelhead}

