% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exRawSthdSmoltData.R
\docType{data}
\name{exRawSthdSmoltData}
\alias{exRawSthdSmoltData}
\title{Example Steelhead Smolt Data from LGTrappingDB}
\format{Data exported from the LGTrappingDB and originally saved as a csv file}
\usage{
exRawSthdSmoltData
}
\description{
This is raw steelhead smolt data 'dumped' out of the LGTrappingDB that was originally saved as a .csv file. No formatting
has been done whatsoever.
}
\keyword{data}
\keyword{smolt}
\keyword{steelhead}

