% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgr2SCOBI.r
\name{lgr2SCOBI}
\alias{lgr2SCOBI}
\title{lgr2SCOBI}
\usage{
lgr2SCOBI(input = NULL, species = "chnk", exportFile = NULL)
}
\arguments{
\item{input}{the name of the csv file containing adult data exported from the LGTrappingDB. Can also accept an \code{object} of a similar
format. See \code{\link[SCOBI]{exRawSthdAdultData}}}

\item{species}{what is the species of the data you are attempting to format for SCOBI. The default option is \code{"chnk"} for
Chinook salmon. Simply set as \code{"sthd"} for steelhead.}

\item{exportFile}{What would you like to name your exported file containing your data formatted for SCOBI analysis? By
default \code{lgr2SCOBI()} exports the formatted data as a csv file which is the preferred format for SCOBI.}
}
\description{
lgr2SCOBI() is a function to convert raw data exported or 'dumped' out of the LGTrappingDB to a format that is ready
to be analyzed by the \link{SCOBI}() function. It is to be used on one species/spawn year combination (e.g., SY2015 steelhead) at a time. It will
write a csv file ready to be used as the \code{adultData} argument of the SCOBI() function.
}
\examples{
lgr2SCOBI(input = exRawSthdAdultData, species = "sthd", exportFile = "SthdScobiInput")

}
\author{
Mike Ackerman
}

