% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgr2SCRAPI.r
\name{lgr2SCRAPI}
\alias{lgr2SCRAPI}
\title{lgr2SCRAPI}
\usage{
lgr2SCRAPI(input = NULL, species = "chnk", exportFile = NULL)
}
\arguments{
\item{input}{the name of the csv file containing smolt data exported from the LGTrappingDB. Can also accept an \code{object} of a similar
format. See \code{\link[SCOBI]{exRawSthdSmoltData}}}

\item{species}{what is the species of the data you are attempting to format for SCRAPI. The default option is \code{"chnk"} for
Chinook salmon. Simply set as \code{"sthd"} for steelhead.}

\item{exportFile}{What would you like to name your exported file containing your data formatted for \code{SCOBI} analysis? By
default \code{lgr2SCRAPI()} exports the formatted data as a csv file which is the preferred format for SCRAPI.}
}
\description{
lgr2SCRAPI() is a function to convert raw data exported or 'dumped' out of the LGTrappingDB to a format that is ready
to be analyzed by the \link{SCRAPI}() function. It is to be used on one species/migratory year combination (e.g., MY2015 steelhead) at
a time. It will write a csv file ready to be used as the \code{smoltData} argument of the SCRAPI() function.
}
\examples{
lgr2SCRAPI(input = exRawSthdSmoltData, species = "sthd", exportFile = "SthdScrapiInput")

}
\author{
Mike Ackerman
}

