% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbtTagRates.R
\docType{data}
\name{pbtTagRates}
\alias{pbtTagRates}
\title{Example PBT Tag Rates Data}
\format{Data is formatted and ready for \code{SCOBI()}. Originally saved as a csv file}
\usage{
pbtTagRates
}
\description{
This is an example dataset of PBT tag rates. This is to be used with \code{pbtTestFish} which is from SY2015 and contains
H individuals with PBT stock assignments
}
\keyword{PBT}
\keyword{adult}
\keyword{rates}
\keyword{steelhead}
\keyword{tag}

