% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbtTestFish.R
\docType{data}
\name{pbtTestFish}
\alias{pbtTestFish}
\title{Steelhead Adult Data with PBT Stock Assignments formatted for SCOBI}
\format{Data is formatted and ready for \code{SCOBI()}. Originally saved as a csv file}
\usage{
pbtTestFish
}
\description{
This is a steelhead adult data set from SY2015 already formatted for \code{SCOBI()}. This dataset contains PBT stock
assignments. PBT stock assignments can be 'expanded' using \code{pbtTagRates}
}
\keyword{PBT}
\keyword{adult}
\keyword{data}
\keyword{steelhead}

