% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCOBI.r
\name{SCOBI}
\alias{SCOBI}
\title{SCOBI}
\usage{
SCOBI(adultData = NULL, windowData = NULL, Run = "output", RTYPE = "W",
  Primary = "GenStock", Secondary = NA, SizeCut = 780, alph = 0.1,
  B = 5000, writeThetas = FALSE, writeOutput = TRUE)
}
\arguments{
\item{adultData}{the name of y csv file containing biological data for adults}

\item{windowData}{the name of the .csv file containing the window count data}

\item{Run}{synopsis of the run being conducted. This will be used as the prefix for all of your output files}

\item{RTYPE}{which rear type (wild [H], hatchery [H], or hatchery unclipped [HNC]) would you like to perform analysis on}

\item{Primary}{the primary category of interest}

\item{Secondary}{the secondary category of interest}

\item{SizeCut}{if Secondary = "LGDFLmm", what FL would you like to use to separate Sm versus Lg fish}

\item{alph}{the alpha used for confidence intervals (e.g., alph = 0.10 -> 90 percent CIs)}

\item{B}{the number of bootstraps to be performed}

\item{writeThetas}{would you like to write the thetas?}

\item{writeOutput}{would you like to write the output files?}
}
\description{
Salmonid Composition Bootstrap Intervals
}
\author{
Kirk Steinhorst and Mike Ackerman
}
\seealso{
\code{\link[MCPAN]{SCSrank}}
}

