% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCOBIpbt.r
\name{SCOBIpbt}
\alias{SCOBIpbt}
\title{SCOBI: Salmonid COmposition Boostrap Intervals}
\usage{
SCOBIpbt(adultData = NULL, windowData = NULL, Run = "output",
  RTYPE = "W", Primary = "GenStock", Secondary = NA, SizeCut = 780,
  alph = 0.1, B = 5000, writeThetas = FALSE, writeOutput = TRUE,
  pbtExpand = FALSE, pbtRates = NULL)
}
\arguments{
\item{adultData}{the .csv file containing the biological data for the adults to be analyzed. The file should contain data for all fish trapped in
a given spawn year and for a given species. The function \code{lgr2SCOBI()} can be used to format data exported from the LGTrappingDB to make it
ready for the \code{SCOBIpbt()} function}

\item{windowData}{the .csv file containing the window count data for the adults to be analyzed. \code{windowData} should contain 3 columns: Strata,
Count, and Collaps}

\item{Run}{synopsis of the run being conducted. \code{Run} will be used as the prefix for all of your output files. \code{Run} should generally
contain the spawn year, rear, species, primary, and (if desired) secondary categories. For example, "sy2015wSthdStockSex"}

\item{RTYPE}{which rear type (wild [H], hatchery [H], or hatchery unclipped [HNC]) would you like to perform the current analysis on}

\item{Primary}{the primary category in the \code{adultData} to be estimated. i.e., after \code{RTYPE} what category would you like to decompose next?}

\item{Secondary}{the secondary category in the \code{adultData} to be estimated. i.e. after \code{Primary} what category would you like to
decompose next? Use \code{Secondary = NA} if no secondary decomposition is desired}

\item{SizeCut}{if Secondary = "LGDFLmm", what FL would you like to use to separate Sm versus Lg fish. For steelhead, set \code{SizeCut = 780}
(the default) to separate A-run and B-run fish. For Chinook, set \code{SizeCut = 570} to separate Jacks and Adults}

\item{alph}{the alpha used for confidence intervals (e.g., \code{alph} = 0.10 results in 90 percent CIs)}

\item{B}{the number of bootstraps to be performed}

\item{writeThetas}{would you like to write the thetas to a .csv? These are the \code{B} bootstrap estimates of the Rear, Primary, and
Secondary (if set) categories.}

\item{writeOutput}{would you like to write output files?}

\item{pbtExpand}{should only be used is \code{RTYPE} is set to "H" or "HNC". If this is an analysis of PBT assigned fish, would you like to expand
the frequencies within each time strata by the PBT tag rates defines in \code{pbtRates}? H stock proportions within each strata will then be
calculated using the expanded frequencies.}

\item{pbtRates}{if \code{pbtExpand = TRUE, SCOBI()} expects a .csv file containing the PBT tag rates for any H stock that occurs in
the \code{adultData}. the \code{pbtRates} should contain 2 columns; the first column should contain the tag rates that correspond exactly
to the H stocks in the \code{adultData}, the second column should contain the PBT tag rate for each H stock}
}
\description{
Perform compositional analyses of adults at Lower Granite Dam.
}
\author{
Kirk Steinhorst and Mike Ackerman
}
\references{
Steinhorst, K., T. Copeland, M. W. Ackerman, W. C. Schrader, E. C. Anderson (In review) Estimates and Confidence Intervals for Run Composition
of Returning Salmonids. Fishery Bulletin.
}
\seealso{
\code{\link[MCPAN]{SCSrank}}
}

