% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCRAPI.r
\name{SCRAPI}
\alias{SCRAPI}
\title{SCRAPI v2.2: The smolt companion to SCOBI}
\usage{
SCRAPI(smoltData = NULL, Dat = "CollectionDate", Rr = "Rear",
  Primary = "GenStock", Secondary = NA, passageData = NULL,
  strat = "Week", dat = "SampleEndDate", tally = "SampleCount",
  samrate = "SampleRate", guidance = "GuidanceEfficiency",
  collaps = "Collapse", Run = "output", RTYPE = "W", REARSTRAT = TRUE,
  alph = 0.1, B = 5000, dateFormat = "\%m/\%d/\%Y")
}
\arguments{
\item{smoltData}{the .csv file containing the biological data for the smolts to be analyzed. The file should contain data for all smolts
trapped in a given migratory year and for a given 'species' (sthd, ch0, or ch1). The function \code{lgr2SCRAPI()} can be used to format
raw smolt data exported from the LGTrappingDB to make it ready for the \code{SCRAPI()} function}

\item{Dat}{speciy the column in \code{smoltData} that contains the sample date for each smolt}

\item{Rr}{specify the column in \code{smoltData} that contains the rear type (W or HNC) for each smolt}

\item{Primary}{specify the primary category in \code{smoltData} to be estimated. i.e., after \code{RTYPE} what category would you like to decompose?}

\item{Secondary}{the secondary category in the \code{smoltData} to be estimated. i.e. after \code{Primary} what category would you like to
decompose next? Use \code{Secondary = NA} if no secondary decomposition is desired}

\item{passageData}{the .csv file containg the smolt passage data. The file should contain the sampling date for each day of the season and
for each date the 1) calendar week, 2) sampling rate, 3) count of smolts in the trap, and 4) the juvenile bypass system guidance efficiency. \code{passageData}
should also contain the collapsing scheme to be used.}

\item{strat}{specify the column in \code{passageData} that contains the calendar weeks}

\item{dat}{specify the column in \code{passageData} that contains the sampling dates}

\item{tally}{specify the column in \code{passageData} containg the daily count of smolts in the trap}

\item{samrate}{specify the column in \code{passageData} containing the daily trap sample rate}

\item{guidance}{specify the column in \code{passageData} containing the daily guidance efficiency estimates}

\item{collaps}{specify the column in \code{passageData} containing the collapsing scheme to be used}

\item{Run}{synopsis of the run being conducted. This will be used as the prefix for all of your output files}

\item{RTYPE}{which rear type (wild [H] or hatchery unclipped [HNC]) would you like to perform analysis on}

\item{REARSTRAT}{would you like pWild calculated on a time-stratified basis (TRUE) or across the entire emigration (FALSE)}

\item{alph}{the alpha used for confidence intervals (e.g., \code{alph} = 0.10 results in 90 percent CIs)}

\item{B}{the desired number of bootstraps}

\item{dateFormat}{What format are the dates in your fish and passage data? The default format used in the LGTrappingDB
is 01/01/1900}
}
\description{
Perform compositional analyses of smolts at Lower Granite Dam.
}
\examples{
SCRAPI(smoltData = sthdScrapiInput, Primary = "GenStock", Secondary = "fwAge",
passageData = sthdSmoltPassData, Run = "sthdSmoltTest", RTYPE = "W", alph = 0.1, B = 200)

SCRAPI(smoltData = chnkScrapiInput, Primary = "GenStock", Secondary = "GenSex",
passageData = sthdSmoltPassData, Run = "ch1SmoltTest", RTYPE = "W", alph = 0.1, B = 200)

}
\author{
Kirk Steinhorst and Mike Ackerman
}

