% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCRAPIv1_0.r
\name{SCRAPIv1.0}
\alias{SCRAPIv1.0}
\title{SCRAPI v1.0}
\usage{
SCRAPIv1.0(smoltData = NULL, Dat = "CollectionDate", Rr = "Rear",
  Primary = "GenStock", Secondary = NA, passageData = NULL,
  strat = "Week", dat = "SampleEndDate", tally = "SampleCount",
  samrate = "SampleRate", guidance = "GuidanceEfficiency",
  collaps = "Collapse", Run = "output", RTYPE = "W", REARSTRAT = TRUE,
  alph = 0.1, B = 5000)
}
\arguments{
\item{smoltData}{the name of the file containing the smolt biological data}

\item{Dat}{the column in smoltData containing the sampling date}

\item{Rr}{column heading containing code for rearing type}

\item{Primary}{This is the primary classification variable for grouping}

\item{Secondary}{Use NA if no secondary composition factor is requested}

\item{passageData}{the name of the file containing the smolt passage data}

\item{strat}{This is the column name for stratification (original weeks)}

\item{dat}{Column heading containing sampling date}

\item{tally}{Column heading containing counts at the dam}

\item{samrate}{Column heading containing trap sampling rate}

\item{guidance}{Column heading containing value of guidance efficiency}

\item{collaps}{column containing the collapsing scheme}

\item{Run}{synopsis of the run being conducted. This will be used as the prefix for all of your output files}

\item{RTYPE}{which rear type (wild [H], hatchery [H], or hatchery unclipped [HNC]) would you like to perform analysis on}

\item{REARSTRAT}{would you like pWild calculated on a time-stratified basis (TRUE) or across the entire emigration (FALSE)}

\item{alph}{the desired alpha level to calculate confidence intervals}

\item{B}{the desired number of bootstraps}
}
\description{
The OUTDATED juvenile companion to SCOBI
}
\author{
Kirk Steinhorst and Mike Ackerman
}

