% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sthdScobiInput.R
\docType{data}
\name{sthdScobiInput}
\alias{sthdScobiInput}
\title{Steelhead Adult Data from LGTrappingDB formatted for SCOBI}
\format{Data is formatted and ready for \code{SCOBI()}. Originally saved as a csv file}
\usage{
sthdScobiInput
}
\description{
This is a steelhead adult data set from SY2015 that has already been formatted for \code{SCOBI()}.
}
\keyword{adult}
\keyword{data}
\keyword{steelhead}

