% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sthdScrapiInput.R
\docType{data}
\name{sthdScrapiInput}
\alias{sthdScrapiInput}
\title{Steelhead Smolt Data from LGTrappingDB formatted for SCRAPI}
\format{Data is formatted and ready for \code{SCRAPI()}. Originally saved as a csv file}
\usage{
sthdScrapiInput
}
\description{
This is a steelhead smolt data set from MY2015 that has already been formatted for \code{SCRAPI()}.
}
\keyword{data}
\keyword{smolt}
\keyword{steelhead}

