% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sthdSmoltPassData.R
\docType{data}
\name{sthdSmoltPassData}
\alias{sthdSmoltPassData}
\title{MY2015 steelhead Smolt Passage Data}
\format{Data is formatted and ready for \code{SCRAPI()}. Originally saved as a csv file}
\usage{
sthdSmoltPassData
}
\description{
This is a steelhead smolt passage data set from MY2015 that has already been formatted for \code{SCRAPI()}.
}
\keyword{data}
\keyword{passage}
\keyword{smolt}
\keyword{steelhead}

