% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sthdWindowCounts.R
\docType{data}
\name{sthdWindowCounts}
\alias{sthdWindowCounts}
\title{SY2015 Steelhead Adult Window Count Data}
\format{Data is formatted and ready for \code{SCOBI()}. Originally saved as a csv file}
\usage{
sthdWindowCounts
}
\description{
This is a steelhead window count data set from SY2015 that has already been formatted for \code{SCOBI()}.
}
\keyword{adult}
\keyword{data}
\keyword{steelhead}
\keyword{window}

